Making ModeratedItem a generic component
